Separate out edit modals
